requirements.txt
adal==1.2.0
alembic==0.8.10
apache-airflow==1.10.0
asn1crypto==0.24.0
Babel==2.6.0
bleach==2.1.2
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
configparser==3.5.0
croniter==0.3.26
cryptography==2.3.1
defusedxml==0.5.0
dill==0.2.8.2
docutils==0.14
Flask==0.12.4
Flask-Admin==1.4.1
Flask-AppBuilder==1.12.2
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Login==0.2.11
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.5
GitPython==2.1.11
google-auth==1.5.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.8.1
kubernetes==8.0.0
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
monotonic==1.5
mysql-connector-python==8.0.13
mysqlclient==1.3.13
numpy==1.15.4
oauthlib==2.1.0
ordereddict==1.1
pandas==0.23.4
pendulum==1.4.4
protobuf==3.6.1
psutil==4.4.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
python-daemon==2.1.2
python-dateutil==2.7.5
python-editor==1.0.3
python-nvd3==0.15.0
python-slugify==1.2.6
python3-openid==3.1.0
pytz==2018.7
pytzdata==2018.7
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
rsa==4.0
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.5
SQLAlchemy==1.1.18
tabulate==0.7.7
tenacity==4.8.0
text-unidecode==1.2
thrift==0.11.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.24.1
webencodings==0.5.1
websocket-client==0.54.0
Werkzeug==0.14.1
WTForms==2.2.1
zope.deprecation==4.3.0
